requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.1.2
django-bootstrap3==14.2.0
django-bootstrap4==2.3.0
django-cors-headers==3.5.0
django-registration==2.4.1
django-taggit==1.3.0
django-tinymce==3.1.0
django-urls==1.1.3
Pillow==8.0.0
python-decouple==3.3
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
